+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix ingredient classication not updating to probition or to specification
 - Make sure totals update when an ingredient is excluded from the formula
 - Fix a bug causing inventory fail when dealing with liters
+- Major logic update in ingredients page
 
 ### Version 3.9
 - Fix a bug showing empty inventory column when exporting a formula